# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TradeShield B2B escrow marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /kyc/submit:
    post:
      operationId: submitKyc
      tags: [kyc]
      summary: Submit Ghana Card details for verification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KycSubmitInput"
      responses:
        "200":
          description: Submission received, now pending admin review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KycSubmitResponse"
        "400":
          description: Validation error or already submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /kyc/status:
    get:
      operationId: getKycStatus
      tags: [kyc]
      summary: Get current user's KYC status
      responses:
        "200":
          description: KYC status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KycStatusResponse"
        "401":
          description: Not logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/kyc/users:
    get:
      operationId: getAdminKycUsers
      tags: [kyc]
      summary: List all users with KYC information (admin only)
      parameters:
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/KycStatus"
          description: Filter by KYC status
      responses:
        "200":
          description: All users with KYC records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminKycUserItem"
        "401":
          description: Not logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/kyc/queue:
    get:
      operationId: getAdminKycQueue
      tags: [kyc]
      summary: List all pending KYC submissions (admin only)
      responses:
        "200":
          description: Pending KYC queue
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminKycQueueItem"
        "401":
          description: Not logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/kyc/{userId}/approve:
    patch:
      operationId: approveKyc
      tags: [kyc]
      summary: Approve a KYC submission (admin only)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User KYC approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KycStatusResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/kyc/{userId}/reject:
    patch:
      operationId: rejectKyc
      tags: [kyc]
      summary: Reject a KYC submission (admin only)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KycRejectInput"
      responses:
        "200":
          description: User KYC rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KycStatusResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    UserRole:
      type: string
      enum: [buyer, supplier, both]
...
    KycStatus:
      type: string
      enum: [none, pending, approved, rejected]
    KycDocType:
      type: string
      enum: [ghana_card_front, ghana_card_back]
    KycDocument:
      type: object
      required: [docType, storageUrl]
      properties:
        docType:
          $ref: "#/components/schemas/KycDocType"
        storageUrl:
          type: string
    KycSubmitInput:
      type: object
      required: [ghanaCardNumber, docUrls]
      properties:
        ghanaCardNumber:
          type: string
          description: Ghana Card number in format GHA-XXXXXXXXX-X
        docUrls:
          type: object
          required: [ghana_card_front, ghana_card_back]
          properties:
            ghana_card_front:
              type: string
              format: uri
            ghana_card_back:
              type: string
              format: uri
    KycSubmitResponse:
      type: object
      required: [submitted]
      properties:
        submitted:
          type: boolean
    KycStatusResponse:
      type: object
      required: [kycStatus]
      properties:
        kycStatus:
          $ref: "#/components/schemas/KycStatus"
        kycRejectionReason:
          type: ["string", "null"]
        ghanaCardNumber:
          type: ["string", "null"]
    KycRejectInput:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
    AdminKycQueueItem:
      type: object
      required: [id, phone, businessName, role, kycSubmittedAt, documents]
      properties:
        id:
          type: integer
        phone:
          type: string
        businessName:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
        kycSubmittedAt:
          type: string
          format: date-time
        ghanaCardNumber:
          type: ["string", "null"]
        documents:
          type: array
          items:
            $ref: "#/components/schemas/KycDocument"
    AdminKycUserItem:
      type: object
      required: [id, phone, businessName, role, kycStatus, documents]
      properties:
        id:
          type: integer
        phone:
          type: string
        businessName:
          type: string
        role:
          $ref: "#/components/schemas/UserRole"
        kycStatus:
          $ref: "#/components/schemas/KycStatus"
        kycSubmittedAt:
          type: ["string", "null"]
          format: date-time
        kycReviewedAt:
          type: ["string", "null"]
          format: date-time
        ghanaCardNumber:
          type: ["string", "null"]
        kycRejectionReason:
          type: ["string", "null"]
        documents:
          type: array
          items:
            $ref: "#/components/schemas/KycDocument"
...